:recycle: Simplify the token generation text and put it into the i18n file
Instead of having separate English and Spanish text encoded in the HTML.
through judicious use of `$.i18n` with the templating mechanism

Did not fix the tokenType just yet since we are going to change it soon anyway.
Removed the tokenType creation from the javascript to make that clear

Also removed unused parts of a failed attempt to handle program vs. study text correctly.
Should also go through and change all the text to HTML instead of having parts,
but that is a separate cleanup.