:goal_net: Set up an error reporting mechanism for invalid study configs
- Create an invalid study div with some internal structure
    - error, error details, and contact for assistance
    - invalid div is hidden by default
- Create a function that will hide the valid div and populate and show the invalid div
- Call the function if there are issues while loading the dynamic config
    - use hardcoded values since the i18n might or might not be loaded correctly

I originally assumed that the error would be generated even if the `i18n` load
was not successful, but that is not the case
https://github.com/wikimedia/jquery.i18n/issues/255

I then tried to force such an error

```
            // i18nLoadPromise always resolves with undefined
            // it explicitly catches failure responses and resolves instead
            // https://github.com/wikimedia/jquery.i18n/blob/6afc05ae64a2051d0a97b143ae90ca76f651874e/src/jquery.i18n.messagestore.js#L31
            // I have filed an issue for this
            // https://github.com/wikimedia/jquery.i18n/issues/255
            // while waiting for it to be discussed, we need to look for a workaround
            // only current way to determine whether or not the load succeeded
            // is to check whether there are messages for the currently loaded language
            if (!$.i18n().messageStore.messages['en']) {
                throw SyntaxError("'en' i18n not found or unparseable");
            }
```

but then realized that the join URL was only dependent on the study config and
not the i18n, so even if the i18n was unsuccessful, we would be able to generate the join URL.
So I removed it (although I have recorded it here for posterity). I also did
not include a catch block in the second round of loading.