use "join" codes instead of generating opcodes on join page
https://github.com/e-mission/e-mission-docs/issues/1076

OPcodes are meant to be used by only one person. For autogen studies, generating them on the join page introduces the possibility that users will misinterpret the instructions and share OPcodes with others.

Instead, we can provide a "join" link which takes users to the app and generates an OPcode there. The "join" link includes the same parameters that the join page can have: study_config, sub_group, and tester.

If viewing the join page on a mobile device, we provide the join link as a button which directly links to the app (assuming it has been installed per instructions).
Otherwise, we provide the link as a QR code.
We also provide the link as text to easily copy and paste.

Several of the functions related to token generation are no longer needed here and are being moved to e-mission-phone. a derivative of _getStudyName is still kept, called getStudyNameFromUrl, consistent with what e-mission-phone will have.